+++ simple_manga_downloader/modules/heavenmanga_org.py
@@ -9,22 +9,15 @@ class Heavenmanga:
         self.site = "heavenmanga.org"
         self.folder = directory
         self.manga_link = link
-        self.base_link = "https://ww2.heavenmanga.org/"
+        self.base_link = "http://ww2.heavenmanga.org/"
 
     def get_chapters(self, title_return):
         '''Gets the list of available chapters
         title_return=True will not create the chapters dict,
         used if only title is needed'''
 
-        try:
-            r = requests.get(self.manga_link, timeout=5)
-        except requests.Timeout:
-            return "Request Timeout"
-        except requests.ConnectionError:
-            return "ConnectionError"
-        if r.status_code != 200:
-            return r.status_code
-
+        r = requests.get(self.manga_link, timeout=5)
+        r.raise_for_status()
         soup = BeautifulSoup(r.text, "html.parser")
 
         self.series_title = soup.find(class_="name bigger").string
@@ -43,14 +36,9 @@ class Heavenmanga:
             else:
                 break
 
-            try:
-                page = requests.get(next_page, timeout=5)
-            except requests.Timeout:
-                return "Request Timeout"
-            if page.status_code != 200:
-                return page.status_code
+            page = requests.get(next_page, timeout=5)
+            page.raise_for_status()
             soup = BeautifulSoup(page.text, "html.parser")
-
             chapter_divs += soup.find_all(class_="two-rows go-border")
 
         self.chapters = {}
@@ -76,25 +64,15 @@ class Heavenmanga:
     def get_info(self, ch):
         '''Gets the needed data abut the chapters from the site'''
         link = self.chapters[ch]["link"]
-        try:
-            r = requests.get(link, timeout=5)
-        except requests.Timeout:
-            return "Request Timeout"
-        if r.status_code != 200:
-            return r.status_code
-
+        r = requests.get(link, timeout=5)
+        r.raise_for_status()
         soup = BeautifulSoup(r.text, "html.parser")
         viewer = soup.find("center")
         pages = [img["src"] for img in viewer.find_all("img")]
 
         # The site has sometimes broken chapters
         # Getting the first one to check if they work
-        try:
-            test = requests.get(pages[0], stream=True, timeout=5)
-        except (requests.ConnectionError, requests.Timeout):
-            return f"Chapter is probably broken\n{link}\n"
-        if not test:
-            return f"Chapter is probably broken\n{link}\n"
-
+        test = requests.get(pages[0], stream=True, timeout=5)
+        test.raise_for_status()
         self.chapters[ch]["pages"] = pages
         return True